[front/agent_builder] - refactor: move SlackIntegration component into SlackTab
 - SlackIntegration component logic has been consolidated within SlackTab to simplify the structure
 - The dedicated SlackIntegration file was removed to centralize related functionality within the settings module
